refactor: extract indexer/infer/it_this.rs (phase 2)
Move the it/this lambda type-inference cluster (16 functions) from
indexer.rs into src/indexer/infer/it_this.rs.  Three helpers used
exclusively by the cluster (line_has_lambda_param, lambda_brace_pos_for_param,
find_last_dot_at_depth_zero) are defined locally; all sig.rs / args.rs
utilities are imported via super::, and Indexer / is_id_char /
find_enclosing_call_name remain in indexer.rs and are reached via
super::super::.

A matching unit-test file (it_this_tests.rs) covers the main public
entry points.  infer/mod.rs re-exports every pub(crate) symbol so the
rest of the crate is unaffected.
